add steel strut that looks like it has a band wrapped around it
meant to be used when something is placed against the side, so that that
thing appears to be mounted to the strut, such as a Technic mains
transformer ("supply converter") mounted near the top of a tall electrical
pole.

Note: uses wallmounted param2 to allow rotation, rather than facedir
mode. I want to allow for Unified Dyes colorization of a derivative of
this node in another mod, and wallmounted mode uses fewer bits for
rotation than facedir, leaving more for color info.